Update on "separate cpp and cuda compilation"
[ghstack-poisoned]
