+++ torchcsprng/csrc/cpu/kernels.cpp
@@ -6,14 +6,13 @@
  * LICENSE file in the root directory of this source tree.
  */
 
-#pragma once
-
-#include "kernels_commons.h"
+#include "../kernels_header.h"
+#include "../kernels_commons.h"
 
 namespace torch {
 namespace csprng {
-namespace cuda {
+namespace cpu {
 
-#include "kernels_decls.h"
+#include "../kernels_body.h"
 
 }}}
