+++ torchcsprng/csrc/kernels_decls.h
@@ -10,44 +10,44 @@
 
 // ==================================================== Random ========================================================
 
-Tensor& random_(Tensor& self, c10::optional<Generator> generator);
+at::Tensor& random_(at::Tensor& self, c10::optional<at::Generator> generator);
 
-Tensor& random_from_to(Tensor& self, int64_t from, optional<int64_t> to, c10::optional<Generator> generator);
+at::Tensor& random_from_to(at::Tensor& self, int64_t from, optional<int64_t> to, c10::optional<at::Generator> generator);
 
-Tensor& random_to(Tensor& self, int64_t to, c10::optional<Generator> generator);
+at::Tensor& random_to(at::Tensor& self, int64_t to, c10::optional<at::Generator> generator);
 
 // ==================================================== Uniform =======================================================
 
-Tensor& uniform_(Tensor& self, double from, double to, c10::optional<Generator> generator);
+at::Tensor& uniform_(at::Tensor& self, double from, double to, c10::optional<at::Generator> generator);
 
 // ==================================================== Normal ========================================================
 
-Tensor& normal_(Tensor& self, double mean, double std, c10::optional<Generator> generator);
+at::Tensor& normal_(at::Tensor& self, double mean, double std, c10::optional<at::Generator> generator);
 
-Tensor& normal_Tensor_float_out(Tensor& output, const Tensor& mean, double std, c10::optional<Generator> gen);
+at::Tensor& normal_Tensor_float_out(at::Tensor& output, const at::Tensor& mean, double std, c10::optional<at::Generator> gen);
 
-Tensor& normal_float_Tensor_out(Tensor& output, double mean, const Tensor& std, c10::optional<Generator> gen);
+at::Tensor& normal_float_Tensor_out(at::Tensor& output, double mean, const at::Tensor& std, c10::optional<at::Generator> gen);
 
-Tensor& normal_Tensor_Tensor_out(Tensor& output, const Tensor& mean, const Tensor& std, c10::optional<Generator> gen);
+at::Tensor& normal_Tensor_Tensor_out(at::Tensor& output, const at::Tensor& mean, const at::Tensor& std, c10::optional<at::Generator> gen);
 
-Tensor normal_Tensor_float(const Tensor& mean, double std, c10::optional<Generator> gen);
+at::Tensor normal_Tensor_float(const at::Tensor& mean, double std, c10::optional<at::Generator> gen);
 
-Tensor normal_float_Tensor(double mean, const Tensor& std, c10::optional<Generator> gen);
+at::Tensor normal_float_Tensor(double mean, const at::Tensor& std, c10::optional<at::Generator> gen);
 
-Tensor normal_Tensor_Tensor(const Tensor& mean, const Tensor& std, c10::optional<Generator> gen);
+at::Tensor normal_Tensor_Tensor(const at::Tensor& mean, const at::Tensor& std, c10::optional<at::Generator> gen);
 
 // ==================================================== Cauchy ========================================================
 
-Tensor& cauchy_(Tensor& self, double median, double sigma, c10::optional<Generator> generator);
+at::Tensor& cauchy_(at::Tensor& self, double median, double sigma, c10::optional<at::Generator> generator);
 
 // ================================================== LogNormal =======================================================
 
-Tensor& log_normal_(Tensor& self, double mean, double std, c10::optional<Generator> gen);
+at::Tensor& log_normal_(at::Tensor& self, double mean, double std, c10::optional<at::Generator> gen);
 
 // ================================================== Geometric =======================================================
 
-Tensor& geometric_(Tensor& self, double p, c10::optional<Generator> gen);
+at::Tensor& geometric_(at::Tensor& self, double p, c10::optional<at::Generator> gen);
 
 // ================================================== Exponential =====================================================
 
-Tensor& exponential_(Tensor& self, double lambda, c10::optional<Generator> gen);
+at::Tensor& exponential_(at::Tensor& self, double lambda, c10::optional<at::Generator> gen);
